django-budgets/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.2
    # via django
astroid==2.12.12
    # via pylint
autopep8==2.0.0
    # via
    #   -r requirements/dev.in
    #   django-silk
build==0.9.0
    # via pip-tools
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via pip-tools
coverage==6.5.0
    # via -r requirements/dev.in
dill==0.3.6
    # via pylint
django==3.2.16
    # via
    #   -r requirements/base.in
    #   django-filter
    #   django-guardian
    #   django-silk
django-environ==0.9.0
    # via -r requirements/base.in
django-filter==22.1
    # via -r requirements/base.in
django-guardian==2.4.0
    # via -r requirements/base.in
django-silk==5.0.2
    # via -r requirements/dev.in
factory-boy==3.2.1
    # via -r requirements/dev.in
faker==15.1.1
    # via factory-boy
flake8==5.0.4
    # via -r requirements/dev.in
gprof2dot==2022.7.29
    # via django-silk
idna==3.4
    # via requests
isort==5.10.1
    # via pylint
jinja2==3.1.2
    # via django-silk
lazy-object-proxy==1.8.0
    # via astroid
markupsafe==2.1.1
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
packaging==21.3
    # via build
pep517==0.13.0
    # via build
pip-tools==6.9.0
    # via -r requirements/dev.in
platformdirs==2.5.2
    # via pylint
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
pyflakes==2.5.0
    # via flake8
pylint==2.15.5
    # via
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.3
    # via -r requirements/dev.in
pylint-plugin-utils==0.7
    # via pylint-django
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   django-silk
    #   faker
pytz==2022.5
    # via django
requests==2.28.1
    # via django-silk
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-silk
tomli==2.0.1
    # via
    #   autopep8
    #   build
    #   pep517
    #   pylint
tomlkit==0.11.6
    # via pylint
urllib3==1.26.12
    # via requests
uwsgi==2.0.21
    # via -r requirements/base.in
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
